+++ packages/encrypted-archive/src/index.node.ts
